cdk/requirements.txt
attrs==20.2.0
aws-cdk.aws-events==1.71.0
aws-cdk.aws-iam==1.71.0
aws-cdk.aws-kms==1.71.0
aws-cdk.aws-s3==1.71.0
aws-cdk.cloud-assembly-schema==1.71.0
aws-cdk.core==1.71.0
aws-cdk.cx-api==1.71.0
aws-cdk.region-info==1.71.0
cattrs==1.1.1
constructs==3.2.7
jsii==1.14.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
